+++ src/main/java/com/example/Tokkit_server/domain/Notice.java
@@ -13,7 +13,7 @@ public class Notice extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long noticeId;
+    private Long id;
 
     private String title;
     private String content;
