+++ src/main/java/com/example/Tokkit_server/service/command/NoticeCommandService.java
@@ -3,7 +3,7 @@ package com.example.Tokkit_server.service.command;
 import com.example.Tokkit_server.apiPayload.code.status.ErrorStatus;
 import com.example.Tokkit_server.apiPayload.exception.GeneralException;
 import com.example.Tokkit_server.domain.Notice;
-import com.example.Tokkit_server.dto.notice.NoticeRequestDto;
+import com.example.Tokkit_server.dto.request.NoticeRequestDto;
 import com.example.Tokkit_server.repository.NoticeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,7 +15,7 @@ public class NoticeCommandService {
 
     public Long createNotice(NoticeRequestDto requestDto) {
         Notice notice = requestDto.to();
-        return noticeRepository.save(notice).getNoticeId();
+        return noticeRepository.save(notice).getId();
     }
 
     public void updateNotice(Long noticeId, NoticeRequestDto requestDto) {
