+++ src/main/java/com/example/Tokkit_server/service/query/NoticeQueryService.java
@@ -3,7 +3,7 @@ package com.example.Tokkit_server.service.query;
 import com.example.Tokkit_server.apiPayload.code.status.ErrorStatus;
 import com.example.Tokkit_server.apiPayload.exception.GeneralException;
 import com.example.Tokkit_server.domain.Notice;
-import com.example.Tokkit_server.dto.notice.NoticeResponseDto;
+import com.example.Tokkit_server.dto.response.NoticeResponseDto;
 import com.example.Tokkit_server.repository.NoticeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
